LPS-129058 Move JournalArticle upgrade to journal-service.
