+++ modules/apps/adaptive-media/adaptive-media-web/src/main/java/com/liferay/adaptive/media/web/internal/upgrade/AMWebUpgrade.java
@@ -16,7 +16,6 @@ package com.liferay.adaptive.media.web.internal.upgrade;
 
 import com.liferay.adaptive.media.web.internal.configuration.AMConfiguration;
 import com.liferay.adaptive.media.web.internal.upgrade.v1_0_0.BlogsEntryDataFileEntryIdUpgradeProcess;
-import com.liferay.adaptive.media.web.internal.upgrade.v1_0_0.JournalArticleDataFileEntryIdUpgradeProcess;
 import com.liferay.blogs.service.BlogsEntryLocalService;
 import com.liferay.portal.configuration.persistence.upgrade.ConfigurationUpgradeStepFactory;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
@@ -45,8 +44,7 @@ public class AMWebUpgrade implements UpgradeStepRegistrator {
 		registry.register(
 			"1.0.1", "1.0.2",
 			new BlogsEntryDataFileEntryIdUpgradeProcess(
-				_blogsEntryLocalService),
-			new JournalArticleDataFileEntryIdUpgradeProcess());
+				_blogsEntryLocalService));
 	}
 
 	@Reference
